src/sampleLambda/requirements.txt
aws-lambda-powertools
pydantic
boto3-stubs